Port `templates delete` command to use cdfrc.json (#223)
* Port `templates delete` command to use cdfrc.json

* Add some spacing

* Use literal tags

* Modify var names as per pr comments

+++ apps/platypus-cdf-cli/src/app/cmds/templates/delete.ts
@@ -1,36 +1,68 @@
 import { TemplatesApiService } from '@platypus/platypus-core';
 import { getCogniteSDKClient } from '../../utils/cogniteSdk';
-import { CommandArgument, CommandArgumentType } from '../../types';
+import { BaseArgs, CommandArgument, CommandArgumentType } from '../../types';
 import { CLICommand } from '@cognite/platypus-cdf-cli/app/common/cli-command';
-import Response from '@cognite/platypus-cdf-cli/app/utils/logger';
+import Response, {
+  DEBUG as _DEBUG,
+} from '@cognite/platypus-cdf-cli/app/utils/logger';
+import { injectRCFile } from '../../common/config';
+import { Arguments } from 'yargs';
+import { unlink } from 'fs/promises';
+import { CONSTANTS } from '../../constants';
+import { join } from 'path';
+import { cwd } from 'process';
 
 export const command = 'delete';
 
 export const desc = 'Delete a template group';
 
+const DEBUG = _DEBUG.extend('templates:delete');
+
+export type DeleteTemplateGroupCommandArgs = BaseArgs & {
+  ['confirm']: boolean;
+};
+
 export const commandArgs = [
   {
-    name: 'externalId',
-    description:
-      'Template group external id to delete (you must have proper permission to execute the same)',
-    prompt: 'Enter the template group external id that you want to delete',
-    type: CommandArgumentType.STRING,
+    name: 'confirm',
+    description: 'Do not ask for confirmation',
+    prompt: 'Are you sure you want to delete this template group?',
+    type: CommandArgumentType.BOOLEAN,
     required: true,
-    example: '$0 templates delete --externalId=some-external-id',
+    example: '$0 templates delete --confirm',
   },
 ] as CommandArgument[];
 
 export class DeleteTemplateGroupCommand extends CLICommand {
-  async execute(args) {
-    const client = getCogniteSDKClient();
-    const templatesApi = new TemplatesApiService(client);
+  @injectRCFile()
+  async execute(args: Arguments<DeleteTemplateGroupCommandArgs>) {
+    try {
+      const id = args.solutionConfig.config.templateId;
+      if (args.confirm) {
+        DEBUG`Deleting template group ${id}`;
 
-    await templatesApi.deleteTemplateGroup({
-      id: args.externalId,
-    });
-    Response.log(
-      `Deleted the template group "${args.externalId}" successfully.`
-    );
+        const client = getCogniteSDKClient();
+        DEBUG`CDF Client Initialized`;
+
+        const templatesApi = new TemplatesApiService(client);
+        DEBUG`Templates API Initialized`;
+
+        await templatesApi.deleteTemplateGroup({
+          id,
+        });
+        DEBUG`Template group deleted successfully`;
+
+        Response.log(`Deleted the template group "${id}" successfully.`);
+
+        unlink(join(cwd(), CONSTANTS.PROJECT_CONFIG_FILE_NAME)).catch((err) => {
+          DEBUG`${err}`;
+          Response.error('Failed to delete the project config file');
+        });
+      }
+    } catch (error) {
+      DEBUG`${error}`;
+      Response.error(`Failed to delete template group: ${error.message}`);
+    }
   }
 }
